To create a project and set properties for a form On the menu bar, choose File > New > Project.
In the Installed Templates list, choose either C# or Visual Basic.

In the list of templates, choose the Windows Forms Application template, name it Math Quiz, and then choose the OK button.

A form that's named Form1.cs or Form1.vb appears, depending on the programming language that you chose.

Choose the form, and then change its Text property to Math Quiz.

The Properties window contains properties for the form.

Change the size of the form to 500 pixels wide by 400 pixels tall.

You can resize the form by dragging its edges until the correct size appears in the lower-left corner of the integrated development environment (IDE). As an alternative, you can change the values of the Size property.

Change the value of the FormBorderStyle property to Fixed3D, and set the MaximizeBox property to False.

These values prevent quiz takers from resizing the form.