To create the time remaining box Add a Label control from the Toolbox, and then set the value of its (Name) property to timeLabel.
This label will become a box in the upper-right corner that shows the number of seconds that remain in the quiz.

Change the AutoSize property to False so that you can resize the box.

Change the BorderStyle property to FixedSingle to draw a line around the box.

Set the Size property to 200, 30.

Move the label to the upper-right corner of the form, where blue spacer lines will appear.

These lines help you align controls on the form.

In the Properties window, choose the Text property, and then choose the Backspace key to clear its value.

Choose the plus sign (+) next to the Font property, and then change the value of the Size property to 15.75.

You can change several font properties, as the following picture shows.

Properties window showing font size

Add another Label control from the Toolbox, and then set its font size to 15.75.

Set the Text property to Time Left.

Move the label so that it lines up just to the left of the timeLabel label.