To add controls for the addition problems Add a Label control from the Toolbox, and then set its Text property to ? (question mark).
Set the AutoSize property to False.

Set the Size property to 60, 50.

Set the font size to 18.

Set the TextAlign property to MiddleCenter.

Set the Location property to 50, 75 to position the control on the form.

Set the (Name) property to plusLeftLabel.

Choose the plusLeftLabel label, and then choose either the Ctrl+C keys or Copy on the Edit menu.

Paste the label three times by choosing either the Ctrl+V keys or Paste on the Edit menu.

Arrange the three new labels so that they are in a row to the right of the plusLeftLabel label.

You can use the spacer lines to space them out and line them up.

Set the value of the second label's Text property to + (plus sign).

Set the value of the third label's (Name) property to plusRightLabel.

Set the value of the fourth label's Text property to = (equal sign).

Add a NumericUpDown control from the Toolbox, set its font size to 18, and set its width to 100.

You'll learn more about this kind of control later.

Line up the NumericUpDown control with the Label controls for the addition problem.

Change the value of the (Name) property for the NumericUpDown control to sum.

You've created the first row, as the following picture shows.

First row of math quiz