To add a start button and set the tab-index order Add a Button control from the Toolbox, and then set its (Name) property to startButton.
Set the Text property to Start the quiz.

Set the font size to 14.

Set the AutoSize property to True, which causes the button to automatically resize to fit the text.

Center the button near the bottom of the form.

Set the value of the TabIndex property for the startButton control to 1.

 Note

The TabIndex property sets the order of the controls when the quiz taker chooses the Tab key. To see how it works, open any dialog box (for example, on the menu bar, choose File > Open), and then choose the Tab key a few times. Watch how your cursor moves from control to control each time that you choose the Tab key. A programmer decided the order when creating that form.

Set the value of the TabIndex property for the NumericUpDown sum control to 2, for the difference control to 3, for the product control to 4, and for the quotient control to 5.

The form should look like the following illustration.

Initial math quiz form

To verify whether the TabIndex property works as you expect, save and run your program by choosing the F5 key, or by choosing Debug > Start Debugging on the menu bar, and then choose the Tab key a few times.